Split up test code into files
